+++ site-main/variables.tf
@@ -2,6 +2,13 @@ variable region {
   default = "us-east-1"
 }
 
+variable project {
+  default = "noproject"
+}
+variable environment {
+  default = "default"
+}
+
 variable domain {}
 variable duplicate-content-penalty-secret {}
 variable deployer {}
@@ -13,4 +20,10 @@ variable routing_rules {
 
 variable not-found-response-path {
   default = "/404.html"
+}
+
+variable "tags" {
+  type        = "map"
+  description = "Optional Tags"
+  default     = {}
 }