+++ site-redirect/main.tf
@@ -50,9 +50,7 @@ resource "aws_s3_bucket" "website_bucket" {
   //    target_prefix = "${var.log_bucket_prefix}"
   //  }
 
-  tags {
-    //    Name = "Staging Website for releasequeue.com"  //    Generator = "http://gohugo.io"  //    Environment = "${var.environment}"
-  }
+  tags = "${merge("${var.tags}",map("Name", "${var.project}-${var.environment}-${var.domain}", "Environment", "${var.environment}", "Project", "${var.project}"))}"
 }
 
 ################################################################################################################
@@ -150,4 +148,7 @@ resource "aws_cloudfront_distribution" "website_cdn" {
   }
 
   aliases = ["${var.domain}"]
+
+  tags = "${merge("${var.tags}",map("Name", "${var.project}-${var.environment}-${var.domain}", "Environment", "${var.environment}", "Project", "${var.project}"))}"
+
 }