+++ site-redirect/variables.tf
@@ -2,8 +2,21 @@ variable region {
   default = "us-east-1"
 }
 
+variable project {
+  default = "noproject"
+}
+variable environment {
+  default = "default"
+}
+
 variable domain {}
 variable target {}
 variable duplicate-content-penalty-secret {}
 variable deployer {}
 variable acm-certificate-arn {}
+
+variable "tags" {
+  type        = "map"
+  description = "Optional Tags"
+  default     = {}
+}